+++ packages/vc-export/src/KiltCredentialV1.ts
@@ -17,8 +17,7 @@ import type {
   IDelegationNode,
 } from '@kiltprotocol/types'
 
-import { CredentialMalformedError } from './verificationUtils.js'
-import * as KiltAttestationProofV1 from './KiltAttestationProofV1.js'
+import { CredentialMalformedError } from './errors.js'
 import { fromGenesisAndRootHash } from './KiltRevocationStatusV1.js'
 import {
   DEFAULT_CREDENTIAL_CONTEXTS,
@@ -327,7 +326,6 @@ export function validateStructure(credential: VerifiableCredential): void {
  * @param input An [[ICredential]] object.
  * @param issuer The issuer of the attestation to this credential (attester).
  * @param chainGenesisHash Genesis hash of the chain against which the credential is verifiable.
- * @param blockHash Hash of any block at which the credential is verifiable (i.e. Attested and not revoked).
  * @param timestamp Timestamp of the block referenced by blockHash in milliseconds since January 1, 1970, UTC (UNIX epoch).
  * @param ctype Optional: The CType object referenced by the [[ICredential]].
  * @returns A KiltCredentialV1 with embedded KiltAttestationProofV1 proof.
@@ -336,10 +334,9 @@ export function fromICredential(
   input: ICredential,
   issuer: DidUri,
   chainGenesisHash: Uint8Array,
-  blockHash: Uint8Array,
   timestamp: number,
   ctype?: ICType
-): VerifiableCredential & Required<Pick<VerifiableCredential, 'proof'>> {
+): VerifiableCredential {
   const {
     legitimations: legitimationsInput,
     delegationId,
@@ -364,75 +361,5 @@ export function fromICredential(
     ...(delegationId && { delegationId }),
   })
 
-  const proof = KiltAttestationProofV1.fromICredential(input, blockHash)
-  return { ...vc, proof }
-}
-
-export type ExpandedContents<
-  T extends VerifiableCredential['credentialSubject']
-> = {
-  [Key in keyof T as `${T['@context']['@vocab']}`]: T[Key]
-} & { '@id': T['id'] }
-
-/**
- * Transforms credentialSubject to an expanded JSON-LD representation.
- *
- * @param credentialSubject The object containing claims about the credentialSubject.
- * @returns The `credentialSubject` where each key is either `@id` or the result of concatenating the `@vocab` with the original key.
- */
-export function jsonLdExpandCredentialSubject<
-  T extends VerifiableCredential['credentialSubject']
->(credentialSubject: T): ExpandedContents<T> {
-  const expandedContents = {}
-  const vocabulary = credentialSubject['@context']['@vocab']
-  Object.entries(credentialSubject).forEach(([key, value]) => {
-    if (key === '@context') return
-    if (key === 'id' || key === 'type') {
-      expandedContents[`@${key}`] = value
-    } else if (key.startsWith(vocabulary) || key.startsWith('@')) {
-      expandedContents[key] = value
-    } else {
-      expandedContents[vocabulary + key] = value
-    }
-  })
-  return expandedContents as ExpandedContents<T>
-}
-
-const delegationIdPattern =
-  /^kilt:delegation\/(?<delegationId>[-a-zA-Z0-9]{1,78})$/
-
-/**
- * Extract the local (i.e., unique within a KILT blockchain network) delegation node identifier from a [[KiltAttesterDelegationV1]] object.
- *
- * @param delegation A [[KiltAttesterDelegationV1]] object.
- * @returns A delegation id.
- */
-export function delegationIdFromAttesterDelegation(
-  delegation: KiltAttesterDelegationV1
-): Uint8Array {
-  if (delegation.type !== KILT_ATTESTER_DELEGATION_V1_TYPE) {
-    throw new Error(`type must be ${KILT_ATTESTER_DELEGATION_V1_TYPE}`)
-  }
-  const match = delegationIdPattern.exec(delegation.id)
-  if (!match || !match.groups?.delegationId)
-    throw new Error(
-      `not a valid id for type ${KILT_ATTESTER_DELEGATION_V1_TYPE}: ${delegation.id}`
-    )
-  return base58Decode(match.groups.delegationId)
-}
-
-/**
- * Extract the local (i.e., unique within a KILT blockchain network) delegation node identifier from a credential's federatedTrustModel entries.
- *
- * @param credential A [[KiltCredentialV1]] type VerifiableCredential.
- * @returns A delegation id or `null` if there is no [[KiltAttesterDelegationV1]] type entry in the federatedTrustModel.
- */
-export function getDelegationNodeIdForCredential(
-  credential: Pick<VerifiableCredential, 'federatedTrustModel'>
-): Uint8Array | null {
-  const delegation = credential.federatedTrustModel?.find(
-    (i): i is KiltAttesterDelegationV1 =>
-      i.type === KILT_ATTESTER_DELEGATION_V1_TYPE
-  )
-  return delegation ? delegationIdFromAttesterDelegation(delegation) : null
+  return vc
 }
